requirements.txt

Flask==3.0.3
pywebview==4.3.3
pyinstaller==6.0.0
jinja-flowbite>=0.3.dev0
